D5-1 Report generation happens in D51 object
